refining the cli api a tiny bit more
trying to start making a line between general cli capabilities and click
specific ones. trying to find a way to bring back lazy loading without
pulling more click details into the plugin apis
