+++ rats-apps/src/python/rats/cli/__init__.py
@@ -1,20 +1,19 @@
 """Uses `rats.annotations` to streamline the creation of CLI commands written with Click."""
 
 from ._annotations import CommandContainer, CommandId, command, group
+from ._click import ClickCommandGroup, ClickCommandMapper
 from ._executable import ClickExecutable, ClickGroup
-from ._groups import ClickGroupPlugin, DeferredCommandGroup, GroupCommands
+from ._groups import ClickGroupPlugin, GroupCommands
 
 __all__ = [
     "CommandId",
     "command",
     "group",
     "CommandContainer",
-    "CommandProvider",
+    "ClickCommandMapper",
     "ClickExecutable",
     "ClickGroupPlugin",
-    "DeferredCommandGroup",
+    "ClickCommandGroup",
     "GroupCommands",
     "ClickGroup",
 ]
-
-from ._provider import CommandProvider
