+++ rats-apps/src/python/rats/cli/_groups.py
@@ -1,24 +1,8 @@
 from abc import abstractmethod
-from typing import Iterator, Protocol, final
+from collections.abc import Iterator
+from typing import Protocol
 
 import click
-
-from ._provider import CommandProvider
-
-
-@final
-class DeferredCommandGroup(click.Group):
-    _provider: CommandProvider
-
-    def __init__(self, name: str, provider: CommandProvider) -> None:
-        super().__init__(name=name)
-        self._provider = provider
-
-    def get_command(self, ctx: click.Context, cmd_name: str) -> click.Command | None:
-        return self._provider.get(cmd_name)
-
-    def list_commands(self, ctx: click.Context) -> list[str]:
-        return list(self._provider.list())
 
 
 class ClickGroupPlugin(Protocol):
